Remove PreallocationSize.
This feature was introduced to improve performance when writing
a file, by specifying up-front how much data will be written.

On Linux, in addition to allocating, we need to update the length
to have a measurable performance improvement.

This change removes the feature for the upcoming release as we
were not able to nail down the appropriate behavior and name.