FileStatus.Unix/Process.Unix: align caching of user identity.
This uses a cached identity on Linux but only when the user has no
permission to change it.

Before, Process.Unix would always cache the identity, and FileStatus
would retrieve it and maintain a local cache for Refresh.

This also adds some additional logic to both Process and FileStatus
which eliminate the need to check the identity in some (common) cases.