[wasm] Fix checking for icudt files against the runtime pack
The files in the appbundle were being compared against the file copied
to the bindir for Wasm?Build.Tests, like
`artifacts/bin/Wasm.Build.Tests/Release/net8.0/browser-wasm`.

But since WBT is not being built to target wasm anymore, these files are
never copied to the bindir. Instead, construct the path to runtime pack
and check against that. This is the correct behavior.