Delete BundleHelper test project (#96693)
The tests that were using the `BundleHelper` to create a single-file app have been moved to the newer infrastructure that uses a pre-built test asset and the live shared framework. The only remaining functions still in use are `GetExtractedFiles` and `AddLongNameContent`. This deletes the `BundleHelper` project and moves those functions to the tests where they are used.