Avoid copy action for root assemblies (#96363)
Changes TrimmerRootAssembly (and the underlying option -a) to
root all members of the assembly, without setting the action to
copy. This allows us to perform branch removal and attribute
removal for the assembly.

This includes changes to MarkStep to ensure that rooted
assemblies are marked visible to reflection. This for example
ensures that parameter names don't get removed.

With the change to MarkEntireType that calls MarkType instead of
Annotations.Mark, we are now reaching MarkType for the dependency
from the copy assembly to the attribute it contains, producing an
unexpected IL2045. Silence this by checking the DependencyKind.

Also fixes a bug in constant method analysis that this uncovered:

BitVector32.CreateMask(int.MinValue) was incorrectly being treated
as returning constant 1. The IL from the beginning of that method:

    IL_0000:  ldarg.0
    IL_0001:  brtrue.s   IL_0005
    IL_0003:  ldc.i4.1
    IL_0004:  ret

The brtrue.s branch was being analyzed as not taken because the
argument was not 1. The branch should be analyzed as taken for
any non-zero argument.