Hardware accelerated implementation of Quaternion multiply (#96624)
* Vector based implementation of Quaternion multiply

I saw on the Discord that Quaternion multiply wasnt yet vectorised, this PR adds that. The non accelerated path is the exact same as before, and the accelerated path appears to return the same results.

* Convert tabs to spaces

* Update src/libraries/System.Private.CoreLib/src/System/Numerics/Quaternion.cs

Copy  paste error!

Co-authored-by: Austin Wise <AustinWise@gmail.com>

* Even faster version

Tanner provided an even quicker version. I've changed it to use GetElementUnsafe as i saw other API's doing the same, and we can reasonably argue these fields to exist.

* Update src/libraries/System.Private.CoreLib/src/System/Numerics/Quaternion.cs

Co-authored-by: Tanner Gooding <tagoo@outlook.com>

* Fix incorrect non-hardware accelerated version

---------

Co-authored-by: Austin Wise <AustinWise@gmail.com>
Co-authored-by: Tanner Gooding <tagoo@outlook.com>