Remove xunit-specific Rosyln analyzer testing packages (#97220)
Microsoft.CodeAnalysis.CSharp.Analyzer.Testing.Xunit brings in an older xunit reference that is not compatible in some cases, we've hit issues with it while upgrading our xunit reference in the past.

As mentioned in https://github.com/dotnet/roslyn-sdk/issues/1099#issuecomment-1723487931 we can use the generic testing package instead.
The only difference is the exception type thrown when an assert fails, e.g. the test throws `InvalidOperationException` instead of `Xunit.Sdk.EqualException` but we still get actual vs. expected exception messages so this is enough for us.